charisma: Reproducible Color Characterization of Digital Images for Biological Studies

Provides a standardized and reproducible framework for characterizing and classifying discrete color classes from digital images of biological organisms. The package automatically determines the presence or absence of 10 human-visible color categories (black, blue, brown, green, grey, orange, purple, red, white, yellow) using a biologically-inspired Color Look-Up Table (CLUT) that partitions HSV color space. Supports both fully automated and semi-automated (interactive) workflows with complete provenance tracking for reproducibility. Pre-processes images using the 'recolorize' package (Weller et al. 2024 doi:10.1111/ele.14378 ) for spatial-color binning, and integrates with 'pavo' (Maia et al. 2019 doi:10.1111/2041-210X.13174 ) for color pattern geometry statistics. Designed for high-throughput analysis and seamless integration with downstream evolutionary analyses.

See also

Author

Maintainer: Shawn Schwartz shawn.t.schwartz@gmail.com (ORCID) [copyright holder]

Authors:

Whitney Tsai wtsai@nhm.org (ORCID)

Elizabeth Karan (ORCID)

Mark Juhn

Allison Shultz ashultz@nhm.org (ORCID)

John McCormack (ORCID)

Thomas Smith (ORCID)

Michael Alfaro michaelalfaro@g.ucla.edu (ORCID)